fix: 변수명 통일 및 기존 변수 deprecated처리, 타입 ㅇouble로 통일

+++ back/tripsok_back/src/main/java/com/tripsok_back/dto/place/PlaceBriefSlimResponseDto.java
@@ -1,6 +1,5 @@
 package com.tripsok_back.dto.place;
 
-import java.math.BigDecimal;
 import java.time.Instant;
 import java.time.ZoneId;
 import java.util.Objects;
@@ -15,8 +14,8 @@ public record PlaceBriefSlimResponseDto(
 	String name,
 	String summary,
 	String type,
-	BigDecimal lat,
-	BigDecimal lng,
+	Double lat,
+	Double lng,
 	String thumbnailUrl,
 	Instant updatedAt
 ) {
@@ -43,8 +42,8 @@ public record PlaceBriefSlimResponseDto(
 			tr != null ? tr.getPlaceName() : null,
 			tr.getSummary(),
 			type,
-			p.getMapY(), // lat
-			p.getMapX(), // lng
+			p.getMapY().doubleValue(), // lat
+			p.getMapX().doubleValue(), // lng
 			thumbnailUrl,
 			p.getUpdatedAt() != null ? p.getUpdatedAt().atZone(ZoneId.of("Asia/Seoul")).toInstant() : null
 		);
@@ -82,8 +81,8 @@ public record PlaceBriefSlimResponseDto(
 			tr.getPlaceName(),
 			tr.getSummary(),
 			type,
-			p.getMapY(),
-			p.getMapX(),
+			p.getMapY().doubleValue(),
+			p.getMapX().doubleValue(),
 			thumb,
 			p.getUpdatedAt().atZone(ZoneId.of("Asia/Seoul")).toInstant()
 		);
@@ -103,8 +102,8 @@ public record PlaceBriefSlimResponseDto(
 			d.getTitle(),
 			d.getSummary(),
 			d.getType(),
-			d.getLat() != null ? BigDecimal.valueOf(d.getLat()) : null,
-			d.getLng() != null ? BigDecimal.valueOf(d.getLng()) : null,
+			d.getLat() != null ? Double.valueOf(d.getLat()) : null,
+			d.getLng() != null ? Double.valueOf(d.getLng()) : null,
 			d.getThumbnailUrl(),
 			d.getUpdatedAt() != null ? Instant.ofEpochMilli(d.getUpdatedAt()) : null
 		);
